refactor: remove builder
The builder pattern is not necessary here as the WASMLayerConfigBuilder was 1:1 with WasmLayerConfig

Tests only check for unit assertions and do not span the usage of tracing. They have been moved to the config module. One test for default values was removed.
